% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseides-package.R
\docType{package}
\name{parseids-package}
\alias{parseids-package}
\alias{parseids}
\title{Parsing Expression Grammar for parsing identifiers}
\description{
Parsing Expression Grammar for parsing identifiers
}
\keyword{package}
