% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dois.R
\name{pid_dois}
\alias{pid_dois}
\title{DOI parsing}
\usage{
pid_dois(x)
}
\arguments{
\item{x}{a character vector of strings with DOIs in them}
}
\description{
DOI parsing
}
\details{
when not found, returns an empty string
}
\examples{
pid_dois("Foo 10.1094/PHYTO-04-17-0144-R hello")
pid_dois("10.1094/PHYTO-04-17-0144-R hello")
pid_dois(" bar 10.1094/PHYTO-04-17-0144-R")
pid_dois("bar 10.1094/PHYTO-04-17-0144-R")
pid_dois("bar bar 10.1094/PHYTO-04-17-0144-R")
pid_dois("bar bar foo bar 10.1094/PHYTO-04-17-0144-R")
pid_dois(" 10.1094/PHYTO-04-17-0144-R ")
pid_dois("10.1094/PHYTO-04-17-0144-R")
pid_dois("10.1094/PHYTO")
pid_dois("10.10944/2")
pid_dois("10.1094")

# many
pid_dois(c("10.1094/PHYTO-04-17-0144-R", "10.1094/PHYTO"))
## sample data, 1000 DOIs
dois
pid_dois(dois)
system.time(replicate(10, pid_dois(dois)))

## split DOIs
# pid_dois_split("10.1094/PHYTO-04-17-0144-R")
# pid_dois_split(c("10.1094/PHYTO-04-17-0144-R", "10.5150/cmcm.2011.086"))
}
