% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_prefixes.R
\name{pid_dois_prefixes}
\alias{pid_dois_prefixes}
\title{Get DOIs prefixes}
\usage{
pid_dois_prefixes(x)
}
\arguments{
\item{x}{(character) a character vector of strings with DOIs}
}
\value{
DOI prefixes (character)
}
\description{
Get DOIs prefixes
}
\examples{
pid_dois_prefixes("10.1094/PHYTO-04-17-0144-R")
pid_dois_prefixes(c("10.1094/PHYTO-04-17-0144-R", "10.5150/cmcm.2011.086"))
pid_dois_prefixes(dois[1:500])

# lots
dois_long <- unlist(replicate(100, dois, simplify = FALSE), TRUE)
length(dois_long)
system.time(pid_dois_prefixes(dois_long))
}
