% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_split.R
\name{pid_dois_split}
\alias{pid_dois_split}
\title{Split DOIs into their pieces}
\usage{
pid_dois_split(x)
}
\arguments{
\item{x}{a character vector of strings with DOIs in them}
}
\description{
Split DOIs into their pieces
}
\details{
when not found, returns an empty string
}
\examples{
## split DOIs
# pid_dois_split("10.1094/PHYTO-04-17-0144-R")
# vapply(dois[1:10], pid_dois_split, "")
# pid_dois_split(c("10.1094/PHYTO-04-17-0144-R", "10.5150/cmcm.2011.086"))
}
\keyword{internal}
